Update ajax.js - more efficient $http retry
Remove redundant creation of deferred object. $timeout already returns a promise that will be resolved with the return value of the function provided as its first parameter; in this case the $http promise.

Set invokeApply parameter of $timeout to false as $http will already cause a $digest cycle to occur.